+++ database/migrations/2024_05_31_231924_add_option_type_to_options_table.php
@@ -6,24 +6,22 @@ use Illuminate\Support\Facades\Schema;
 
 class AddOptionTypeToOptionsTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::table('options', function (Blueprint $table) {
-            $table->unsignedBigInteger('option_type_id')->after('question_id');
-            $table->foreign('option_type_id')->references('id')->on('option_types')->onDelete('cascade');
+            if (Schema::hasTable('options') && !Schema::hasColumn('options', 'option_type_id')) {
+                $table->foreignId('option_type_id')->after('question_id')->constrained()->onDelete('cascade');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('options', function (Blueprint $table) {
-            $table->dropForeign(['option_type_id']);
-            $table->dropColumn('option_type_id');
+            if (Schema::hasColumn('options', 'option_type_id')) {
+                $table->dropForeign(['option_type_id']);
+                $table->dropColumn('option_type_id');
+            }
         });
     }
 }
